feat(new): new tests and restructure
